Exercise 1b.4
Write a Python function is_leap_year that take as input the parameter
year and returns True if year (an integer) is a leap year according to
the Gregorian calendar and False otherwise. The Wikipedia entry for
leap yearscontains a simple algorithmic rule for determining whether a
year is a leap year. Your main task will be to translate this rule into
Python.