Exercise 1b.7
Write a Python function print_digits that takes an integer number in
the range [0,100) and prints the message "The tens digit is %, and the
ones digit is %." where the percents should be replaced with the
appropriate values. The function should include an error check for the
case when number is negative or greater than or equal to 100. In those
cases, the function should instead print "Error: Input is not a
two-digit number.".