Exercise 2a.2
Given the program template below, write a Python function set_goodbye()
that updates a global variable message with the value "Goodbye" and
prints the value of this global variable to the console. Note that the
existing global variable message has its original value "Hello"
modified to "Goodbye" during the call to set_goodbye().