Exercise 2a.3
Challenge: Given the program template below, implement four functions
that manipulate a global variable count as follows. The function
reset() sets the value of count to be zero, the function increment()
adds one to count, the function decrement() subtracts one from count,
and the function print_count() that prints the value of count to the
console.