Exercise 2b.3
Challenge: Given the program template below, implement four buttons
that manipulate a global variable count as follows. The function
reset() sets the value of count to be zero, the function increment()
adds one to count, the function decrement() subtracts one from count,
and the function print_count() prints the value of count to the console.