Pig Latin
Write a program allows a user to enter a word in an input field,
translates that word into Pig Latin and prints this translation in the
console. For the sake of modularity, we suggest that you build a helper
function that handles all of the details of translating a word to Pig
Latin (see the practice exercises for logic and conditionals) . The
provided template includes the operations for extracting the first
letter and rest of the input word in the partial definition of this
function.