RPSLS
Add an interactive user interface for your implementation of
"Rock-paper-scissors-lizard-Spock". Create an input field that takes a
player's guess, generates a random computer guess, and prints out the
player and computer choices as well as who won in the console. Make
sure that your program checks for and correctly responds to bad input.