Mini-project #2 - "Guess the Number!"
One of the simplest two-player games is “Guess the number”. The first
player thinks of a secret number in some known range while the second
player attempts to guess the number. After each guess, the first player
answers either “Higher”, “Lower” or “Correct!” depending on whether the
secret number is higher, lower or equal to the guess. In this project,
you will build a simple interactive program in Python where the
computer will take the role of the first player while you play as the
second player.