Week 3a.5
Challenge:Complete the program template below and add two buttons that
control the radius of a white ball centered in the middle of the
canvas. Clicking the “Increase radius” button should increase the
radius of the ball. Clicking the “Decrease radius” button should
decrease the radius of the ball, except that the ball radius should
always be positive.